% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/air polution.R
\name{air_polution_forecast}
\alias{air_polution_forecast}
\title{Air polution forecast data}
\usage{
air_polution_forecast(
  lat,
  lon,
  api_key = Sys.getenv("weather_key"),
  lang = "en",
  units = "standard",
  tibble_format = FALSE
)
}
\arguments{
\item{lat}{Latitude in decimals}

\item{lon}{Longitude in decimals}

\item{api_key}{API key from OpenWeather (not necessary if an API key was previosly set)}

\item{lang}{Supported languaje}

\item{units}{Units of measurement. The type of unit of measure can be "standard" (default), "metric" and "imperial"}

\item{tibble_format}{Logical value. If it is TRUE returns the request on tibble format}
}
\value{

}
\description{
Air polution forecast data
}
