% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/air polution.R
\name{air_polution_historical}
\alias{air_polution_historical}
\title{Air polution historical data}
\usage{
air_polution_historical(
  lat,
  lon,
  api_key = Sys.getenv("weather_key"),
  lang = "en",
  units = "standard",
  tibble_format = FALSE,
  start_date = "2021-01-01",
  end_date = "2021-01-02"
)
}
\arguments{
\item{lat}{Latitude in decimals}

\item{lon}{Longitude in decimals}

\item{api_key}{API key from OpenWeather (not necessary if an API key was previosly set)}

\item{lang}{Supported languaje}

\item{units}{Units of measurement. The type of unit of measure can be "standard" (default), "metric" and "imperial"}

\item{tibble_format}{Logical value. If it is TRUE returns the request on tibble format}

\item{start_date}{Initial date in format YYYY/MM/DD.}

\item{end_date}{Final date in format YYYY/MM/DD.}
}
\value{

}
\description{
Air polution historical data
}
