% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-cities.R
\docType{data}
\name{cities}
\alias{cities}
\title{Cities available to obtain their weather data}
\format{
A data frame with 209579 rows and 5 variables:
\describe{
\item{id}{Row ID}
\item{name}{City name}
\item{country}{Country name}
\item{coord.lon}{Longitude}
\item{coord.lat}{Latitude}
}
}
\usage{
cities
}
\description{
Data obtained directly from <http://bulk.openweathermap.org/sample/>
}
\examples{
cities
}
\keyword{datasets}
