% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generals.R
\name{get_coordinates}
\alias{get_coordinates}
\title{Generic function to obtain coordinates from country/state}
\usage{
get_coordinates(
  country,
  state,
  api_key = Sys.getenv("weather_key"),
  only_coordinates = FALSE
)
}
\arguments{
\item{country}{String with country name}

\item{state}{String with state name}

\item{api_key}{API key from OpenWeather (not necessary if an API key was previosly set)}

\item{only_coordinates}{Logical value. If it is TRUE it will only return the coordinates on a data frame. FALSE by default.}
}
\value{

}
\description{
Generic function to obtain coordinates from country/state
}
