% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generals.R
\name{set_weather_key}
\alias{set_weather_key}
\title{Set weather API key}
\usage{
set_weather_key(api_key)
}
\arguments{
\item{api_key}{API key from OpenWeather}
}
\value{

}
\description{
Set weather API key
}
