% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/current weather data.R
\name{weather_actual_req}
\alias{weather_actual_req}
\title{Actual weather data request}
\usage{
weather_actual_req(
  lat,
  lon,
  api_key = Sys.getenv("weather_key"),
  lang = "en",
  units = "standard",
  tibble_format = FALSE
)
}
\arguments{
\item{lat}{Latitude in decimals}

\item{lon}{Longitude in decimals}

\item{api_key}{API key from OpenWeather (not necessary if an API key was previosly set)}

\item{lang}{Supported languaje}

\item{units}{Units of measurement. The type of unit of measure can be "standard" (default), "metric" and "imperial"}

\item{tibble_format}{Logical value. If it is TRUE returns the request on tibble format}
}
\value{

}
\description{
Actual weather data request
}
