% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{weather_forecast_req}
\alias{weather_forecast_req}
\title{Weather forecast data request (5 days. Data every 3 hours)}
\usage{
weather_forecast_req(
  lat,
  lon,
  api_key = Sys.getenv("weather_key"),
  lang = "en",
  units = "standard",
  tibble_format = TRUE
)
}
\arguments{
\item{lat}{Latitude in decimals}

\item{lon}{Longitude in decimals}

\item{api_key}{API key from OpenWeather (not necessary if an API key was previosly set)}

\item{lang}{Supported languaje}

\item{units}{Units of measurement. The type of unit of measure can be "standard" (default), "metric" and "imperial"}

\item{tibble_format}{Logical value. If it is TRUE returns the request on tibble format}
}
\value{

}
\description{
Weather forecast data request (5 days. Data every 3 hours)
}
