% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast-pro.R
\name{weather_forecast_req_hourly}
\alias{weather_forecast_req_hourly}
\title{Weather forecast data request (4 days with hourly data)}
\usage{
weather_forecast_req_hourly(
  lat,
  lon,
  api_key = Sys.getenv("weather_key"),
  lang = "en",
  units = "standard",
  tibble_format = TRUE
)

weather_forecast_req_hourly(
  lat,
  lon,
  api_key = Sys.getenv("weather_key"),
  lang = "en",
  units = "standard",
  tibble_format = TRUE
)
}
\arguments{
\item{lat}{Latitude in decimals}

\item{lon}{Longitude in decimals}

\item{api_key}{API key from OpenWeather (not necessary if an API key was previosly set)}

\item{lang}{Supported languaje}

\item{units}{Units of measurement. The type of unit of measure can be "standard" (default), "metric" and "imperial"}

\item{tibble_format}{Logical value. If it is TRUE returns the request on tibble format}
}
\value{



}
\description{
Weather forecast data request (4 days with hourly data)

Weather forecast data request (Daily data by 16 days)
}
\details{
Note that this function requires a professional account in OpenWeather

Note that this function requires a professional account in OpenWeather
}
