% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generals.R
\name{weather_historic_req}
\alias{weather_historic_req}
\title{Historic weather data request (A single observation of the past)}
\usage{
weather_historic_req(
  lat,
  lon,
  api_key = Sys.getenv("weather_key"),
  lang = "en",
  units = "standard",
  date = "2023-01-01",
  tibble_format = FALSE
)
}
\arguments{
\item{lat}{Latitude in decimals}

\item{lon}{Longitude in decimals}

\item{api_key}{API key from OpenWeather (not necessary if an API key was previosly set)}

\item{lang}{Supported languaje}

\item{units}{Units of measurement. The type of unit of measure can be "standard" (default), "metric" and "imperial"}

\item{date}{Date in format YYYY/MM/DD. The most current date must be at least 5 days before the current day}

\item{tibble_format}{Logical value. If it is TRUE returns the request on tibble format}
}
\value{

}
\description{
Historic weather data request (A single observation of the past)
}
