% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/current weather data.R
\name{weather_state_actual}
\alias{weather_state_actual}
\title{Actual weather data request by name}
\usage{
weather_state_actual(
  api_key = Sys.getenv("weather_key"),
  lang = "en",
  units = "standard",
  name,
  tibble_format = FALSE
)
}
\arguments{
\item{api_key}{API key from OpenWeather (not necessary if an API key was previosly set)}

\item{lang}{Supported languaje}

\item{units}{Units of measurement. The type of unit of measure can be "standard" (default), "metric" and "imperial"}

\item{name}{String with the name of a state (only works for the USA) or country}

\item{tibble_format}{Logical value. If it is TRUE returns the request on tibble format}
}
\value{

}
\description{
Actual weather data request by name
}
