% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{weather_state_forecast}
\alias{weather_state_forecast}
\title{Weather forecast data request by name (5 days. Data every 3 hours)}
\usage{
weather_state_forecast(
  api_key = Sys.getenv("weather_key"),
  lang = "en",
  units = "standard",
  name,
  tibble_format = TRUE
)
}
\arguments{
\item{api_key}{API key from OpenWeather (not necessary if an API key was previosly set)}

\item{lang}{Supported languaje}

\item{units}{Units of measurement. The type of unit of measure can be "standard" (default), "metric" and "imperial"}

\item{name}{String with the name of a state (only works for the USA) or country}

\item{tibble_format}{Logical value. If it is TRUE returns the request on tibble format}
}
\value{

}
\description{
Weather forecast data request by name (5 days. Data every 3 hours)
}
