% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast-pro.R
\name{weather_state_forecast_hourly}
\alias{weather_state_forecast_hourly}
\title{Weather forecast data request by name (4 days with hourly data)}
\usage{
weather_state_forecast_hourly(
  api_key = Sys.getenv("weather_key"),
  lang = "en",
  units = "standard",
  name,
  tibble_format = TRUE
)

weather_state_forecast_hourly(
  api_key = Sys.getenv("weather_key"),
  lang = "en",
  units = "standard",
  name,
  tibble_format = TRUE
)
}
\arguments{
\item{api_key}{API key from OpenWeather (not necessary if an API key was previosly set)}

\item{lang}{Supported languaje}

\item{units}{Units of measurement. The type of unit of measure can be "standard" (default), "metric" and "imperial"}

\item{name}{String with the name of a state (only works for the USA) or country}

\item{tibble_format}{Logical value. If it is TRUE returns the request on tibble format}
}
\value{



}
\description{
Weather forecast data request by name (4 days with hourly data)

Weather forecast data request by name (Daily data by 16 days)
}
\details{
Note that this function requires a professional account in OpenWeather

Note that this function requires a professional account in OpenWeather
}
