% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generals.R
\name{weather_state_historic}
\alias{weather_state_historic}
\title{Historic weather data request (A single observation of the past)}
\usage{
weather_state_historic(
  api_key = Sys.getenv("weather_key"),
  lang = "en",
  units = "standard",
  name,
  date = "2023-01-01",
  tibble_format = FALSE
)
}
\arguments{
\item{api_key}{API key from OpenWeather (not necessary if an API key was previosly set)}

\item{lang}{Supported languaje}

\item{units}{Units of measurement. The type of unit of measure can be "standard" (default), "metric" and "imperial"}

\item{name}{String with the name of a state (only works for the USA) or country}

\item{date}{Date in format YYYY/MM/DD. The most current date must be at least 5 days before the current day}

\item{tibble_format}{Logical value. If it is TRUE returns the request on tibble format}
}
\value{

}
\description{
Historic weather data request (A single observation of the past)
}
